Break DataObject classes into separate files
