+++ src/Microsoft.DotNet.Wpf/src/PresentationCore/PresentationCore.csproj
@@ -319,6 +319,11 @@
     <Compile Include="System\IO\Packaging\PackWebRequest.cs" />
     <Compile Include="System\IO\Packaging\PackWebRequestFactory.cs" />
     <Compile Include="System\IO\Packaging\PackWebResponse.cs" />
+    <Compile Include="System\Windows\DataObject.DataStore.cs" />
+    <Compile Include="System\Windows\DataObject.FormatEnumerator.cs" />
+    <Compile Include="System\Windows\DataObject.OleConverter.cs" />
+    <Compile Include="System\Windows\DataObject.OleConverter.RestrictedTypeDeserializationException.cs" />
+    <Compile Include="System\Windows\DataObject.OleConverter.TypeRestrictingSerializationBinder.cs" />
     <Compile Include="System\Windows\Generated\ContentElement.cs" />
     <Compile Include="System\Windows\Generated\TextDecoration.cs" />
     <Compile Include="System\Windows\Generated\TextDecorationCollection.cs" />
